Cleaned up commands.go file. Moved most commands to their own separate file. Since since map and mapb are so closely tied, placed them in the same file. Moved struct definitions to own files.
